created chat cli, improved simple generation script.
