+++ inference/chat_cli.py
@@ -5,174 +5,13 @@ Code based on: https://github.com/yanghaojin/FastChat/blob/greenbit/fastchat/ser
 
 import argparse
 import os
-import re
 import sys
+import warnings
+warnings.filterwarnings("ignore", category=UserWarning, module='torch.nn.modules.module')
 
-from prompt_toolkit import PromptSession
-from prompt_toolkit.auto_suggest import AutoSuggestFromHistory
-from prompt_toolkit.completion import WordCompleter
-from prompt_toolkit.history import InMemoryHistory
-from prompt_toolkit.key_binding import KeyBindings
-from rich.console import Console
-from rich.live import Live
-from rich.markdown import Markdown
+import torch
 
-from .chat_base import ChatIO, chat_loop
-from .utils import str_to_torch_dtype, is_chat_model
-
-DEFAULT_SEQLEN = 2048
-
-
-class SimpleChatIO(ChatIO):
-    def __init__(self, multiline: bool = False):
-        self._multiline = multiline
-
-    def prompt_for_input(self, role) -> str:
-        if not self._multiline:
-            return input(f"{role}: ")
-
-        prompt_data = []
-        line = input(f"{role} [ctrl-d/z on empty line to end]: ")
-        while True:
-            prompt_data.append(line.strip())
-            try:
-                line = input()
-            except EOFError as e:
-                break
-        return "\n".join(prompt_data)
-
-    def prompt_for_output(self, role: str):
-        print(f"{role}: ", end="", flush=True)
-
-    def stream_output(self, output_stream):
-        pre = 0
-        for outputs in output_stream:
-            output_text = outputs["text"]
-            output_text = output_text.strip().split(" ")
-            now = len(output_text) - 1
-            if now > pre:
-                print(" ".join(output_text[pre:now]), end=" ", flush=True)
-                pre = now
-        print(" ".join(output_text[pre:]), flush=True)
-        return " ".join(output_text)
-
-    def print_output(self, text: str):
-        print(text)
-
-
-class RichChatIO(ChatIO):
-    bindings = KeyBindings()
-
-    @bindings.add("escape", "enter")
-    def _(event):
-        event.app.current_buffer.newline()
-
-    def __init__(self, multiline: bool = False, mouse: bool = False):
-        self._prompt_session = PromptSession(history=InMemoryHistory())
-        self._completer = WordCompleter(
-            words=["!!exit", "!!reset", "!!remove", "!!regen", "!!save", "!!load"],
-            pattern=re.compile("$"),
-        )
-        self._console = Console()
-        self._multiline = multiline
-        self._mouse = mouse
-
-    def prompt_for_input(self, role) -> str:
-        self._console.print(f"[bold]{role}:")
-        # TODO(suquark): multiline input has some issues. fix it later.
-        prompt_input = self._prompt_session.prompt(
-            completer=self._completer,
-            multiline=False,
-            mouse_support=self._mouse,
-            auto_suggest=AutoSuggestFromHistory(),
-            key_bindings=self.bindings if self._multiline else None,
-        )
-        self._console.print()
-        return prompt_input
-
-    def prompt_for_output(self, role: str):
-        self._console.print(f"[bold]{role.replace('/', '|')}:")
-
-    def stream_output(self, output_stream):
-        """Stream output from a role."""
-        # TODO(suquark): the console flickers when there is a code block
-        #  above it. We need to cut off "live" when a code block is done.
-
-        # Create a Live context for updating the console output
-        with Live(console=self._console, refresh_per_second=4) as live:
-            # Read lines from the stream
-            for outputs in output_stream:
-                if not outputs:
-                    continue
-                text = outputs["text"]
-                # Render the accumulated text as Markdown
-                # NOTE: this is a workaround for the rendering "unstandard markdown"
-                #  in rich. The chatbots output treat "\n" as a new line for
-                #  better compatibility with real-world text. However, rendering
-                #  in markdown would break the format. It is because standard markdown
-                #  treat a single "\n" in normal text as a space.
-                #  Our workaround is adding two spaces at the end of each line.
-                #  This is not a perfect solution, as it would
-                #  introduce trailing spaces (only) in code block, but it works well
-                #  especially for console output, because in general the console does not
-                #  care about trailing spaces.
-                lines = []
-                for line in text.splitlines():
-                    lines.append(line)
-                    if line.startswith("```"):
-                        # Code block marker - do not add trailing spaces, as it would
-                        #  break the syntax highlighting
-                        lines.append("\n")
-                    else:
-                        lines.append("  \n")
-                markdown = Markdown("".join(lines))
-                # Update the Live console output
-                live.update(markdown)
-        self._console.print()
-        return text
-
-    def print_output(self, text: str):
-        self.stream_output([{"text": text}])
-
-
-class ProgrammaticChatIO(ChatIO):
-    def prompt_for_input(self, role) -> str:
-        contents = ""
-        # `end_sequence` signals the end of a message. It is unlikely to occur in
-        #  message content.
-        end_sequence = " __END_OF_A_MESSAGE_47582648__\n"
-        len_end = len(end_sequence)
-        while True:
-            if len(contents) >= len_end:
-                last_chars = contents[-len_end:]
-                if last_chars == end_sequence:
-                    break
-            try:
-                char = sys.stdin.read(1)
-                contents = contents + char
-            except EOFError:
-                continue
-        contents = contents[:-len_end]
-        print(f"[!OP:{role}]: {contents}", flush=True)
-        return contents
-
-    def prompt_for_output(self, role: str):
-        print(f"[!OP:{role}]: ", end="", flush=True)
-
-    def stream_output(self, output_stream):
-        pre = 0
-        for outputs in output_stream:
-            output_text = outputs["text"]
-            output_text = output_text.strip().split(" ")
-            now = len(output_text) - 1
-            if now > pre:
-                print(" ".join(output_text[pre:now]), end=" ", flush=True)
-                pre = now
-        print(" ".join(output_text[pre:]), flush=True)
-        return " ".join(output_text)
-
-    def print_output(self, text: str):
-        print(text)
+from utils import str_to_torch_dtype, is_chat_model
 
 
 def main(args):
@@ -182,13 +21,21 @@ def main(args):
             raise ValueError(
                 f"Larger --num-gpus ({args.num_gpus}) than --gpus {args.gpus}!"
             )
+        # NOTE that we need to set this before any other cuda operations. Otherwise will not work.
         os.environ["CUDA_VISIBLE_DEVICES"] = args.gpus
+
+    if not torch.cuda.is_available():
+        raise Exception("Warning: CUDA is needed to run the model.")
+
+    try:
+        from chat_base import chat_loop, SimpleChatIO, RichChatIO
+    except Exception:
+        raise Exception("Error occurred when import chat loop, ChatIO classes.")
+
     if args.style == "simple":
         chatio = SimpleChatIO(args.multiline)
     elif args.style == "rich":
         chatio = RichChatIO(args.multiline, args.mouse)
-    elif args.style == "programmatic":
-        chatio = ProgrammaticChatIO()
     else:
         raise ValueError(f"Invalid style for console: {args.style}")
 
@@ -248,21 +95,16 @@ if __name__ == "__main__":
     parser.add_argument(
         "--gpus",
         type=str,
-        default=None,
+        default='0',
         help="A single GPU like 1 or multiple GPUs like 0,2",
     )
     parser.add_argument("--num-gpus", type=int, default=1)
-    parser.add_argument(
-        "--max-gpu-memory",
-        type=str,
-        help="The maximum memory per GPU for storing model weights. Use a string like '13Gib'",
-    )
     parser.add_argument(
         "--dtype",
         type=str,
         choices=["float32", "float16"],
         help="Override the default dtype. If not set, it will use float16 on GPU and float32 on CPU.",
-        default=None,
+        default="float16",
     )
     parser.add_argument(
         "--conv-template", type=str, default=None, help="Conversation prompt template."
@@ -273,12 +115,12 @@ if __name__ == "__main__":
     parser.add_argument("--temperature", type=float, default=0.7)
     parser.add_argument("--repetition_penalty", type=float, default=1.0)
     parser.add_argument("--max-new-tokens", type=int, default=256)
-    parser.add_argument("--no-history", action="store_true")
+    parser.add_argument("--no-history", action="store_true", help="Disables chat history.")
     parser.add_argument(
         "--style",
         type=str,
-        default="simple",
-        choices=["simple", "rich", "programmatic"],
+        default="rich",
+        choices=["simple", "rich"],
         help="Display style.",
     )
     parser.add_argument(
@@ -313,7 +155,12 @@ if __name__ == "__main__":
         help="Enable trusting remote code",
     )
     parser.add_argument(
-        "--seqlen", type=int, default=DEFAULT_SEQLEN, help="Sequence length"
+        "--seqlen", type=int, default=2048, help="Sequence length"
+    )
+    parser.add_argument(
+        "--use-flash-attention-2",
+        action="store_true",
+        help="Enable using flash attention v2",
     )
     args = parser.parse_args()
     main(args)
